flexible/pubsub/src/main/java/com/example/flexible/pubsub/PubSubPublish.java
/**
 * Copyright 2016 Google Inc. All Rights Reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.example.flexible.pubsub;

import com.google.cloud.pubsub.Message;
import com.google.cloud.pubsub.PubSub;
import com.google.cloud.pubsub.PubSubOptions;
import com.google.cloud.pubsub.Topic;

import java.io.IOException;
import javax.servlet.ServletException;
import javax.servlet.annotation.WebServlet;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;

@WebServlet(name = "PubSubPublish", value = "/publish")
public class PubSubPublish extends HttpServlet {
  @Override
  public void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException,
      ServletException {
    // Load environment variables
    final String topicName = System.getenv("PUBSUB_TOPIC");
    final String payload = req.getParameter("payload");

    // Create a PubSub Service and get Topic
    PubSub pubsub = PubSubOptions.getDefaultInstance().getService();
    Topic topic = pubsub.getTopic(topicName);

    // Publish Message
    topic.publish(Message.of(payload));

    // redirect "/"
    resp.sendRedirect("/");
  }
}
